LayoutTests/platform/ios-simulator/fast/dom/Window/window-property-invalid-characters-ignored-expected.txt

Tests that invalid characters are ignored after a window property value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL w.document.width should be 123. Was 800.
FAIL w.document.height should be 123. Was 600.
PASS w !== window is true
PASS successfullyParsed is true

TEST COMPLETE
